feat(1): primitive phase-1,phase-2
This commit is for 1_dots.

This adds some primitive code that gets us moving between phase-1 (blank
canvas) to phase-2 (dot matrix). There are no transitions/animations and
a fixed interval is used.

Furthermore the canvas is a fixed size for simplicity at this point.

A next step is drawing a line. I think a new data structure should be
introduced to make that task easier. Perhaps a graph or linked list data
structure would work.

If we remove the constraint of crossing paths we've already crossed, it
might be easier at a first pass (and perhaps the drawings could become
more interesting as well).