feat(1): generatePath; drawShape state
This is an admittedly huge chunk of change for the 1_dots project.

It's likely best not to make sense of this commit relative to the
previous one, as I added tests and created new dir structures etc. None
of these structures were clear until some code was written; it just
happened naturally.

The main thing to focus on here is model/generatePath and the
controller/controller.ts files.